formulario vista y template comentario

+++ videorsdb/views.py
@@ -2,9 +2,9 @@ from django.shortcuts import render, redirect
 from django.http import HttpResponseRedirect
 from django.contrib.auth.models import User
 from django.contrib import messages
-from .forms import RegistroUserForm, login_user, subirvideo, crear_canal
+from .forms import RegistroUserForm, login_user, subirvideo, crear_canal, comentario_form
 from django.contrib.auth import authenticate, login, logout
-from .models import UploadVideo, Tagvideo, Perfil, Canal, Subcriptores
+from .models import UploadVideo, Tagvideo, Perfil, Canal, Subcriptores, Comentario
 from .funciones import calcular_codigo, codigo_canal
 from django.http import Http404
 # Create your views here.
@@ -102,28 +102,40 @@ def up_video(request):
     return render(request, 'up_video.html', {'form':form})
 
 def watchvideo(request):
-      vervideo = 'vervideo.html'
-      x = request.GET.get('v','')
-      q = UploadVideo.objects.get(cod_video=x)
-      if request.method == 'POST':
-        if request.user.is_authenticated():
-          u = request.user
-          y = User.objects.get(username=u)
-          r = Canal.objects.get(id_canal=q.id_c.id_canal)
-          try :
-             z = Subcriptores.objects.get(id_u=y.id, id_c=r)
-             z.delete()
-             return render(request, vervideo, {'q':q})
-          except Subcriptores.DoesNotExist:
-             r = Canal.objects.get(id_canal=q.id_c.id_canal)
-             s = Subcriptores(id_c=r, id_u=u)
-             s.save()
+	u = request.user
+	vervideo = 'vervideo.html'
+	x = request.GET.get('v','')
+	q = UploadVideo.objects.get(cod_video=x)
+	c = Comentario.objects.filter(id_v=q.id_video)
+	form = comentario_form()
+	if request.method == 'POST':
+		if request.user.is_authenticated():
+			if 'boton_form_1' in request.POST :
+				form = comentario_form(request.POST)
+				if form.is_valid():
+					comentario = request.POST['comentario']
+					coments = Comentario(comentario_text = comentario, id_v=q, id_user=u)
+					coments.save()
+					return HttpResponseRedirect('/')
+				else :
+					return render(request, vervideo, {'q':q, 'form':form,'c':c})
+			if 'boton_form_2' in request.POST :
+				y = User.objects.get(username=u)
+				r = Canal.objects.get(id_canal=q.id_c.id_canal)
+				try :
+					z = Subcriptores.objects.get(id_u=y.id, id_c=r)
+					z.delete()
+ 					return render(request, vervideo, {'q':q, 'form':form,'c':c})
+				except Subcriptores.DoesNotExist:
+					r = Canal.objects.get(id_canal=q.id_c.id_canal)
+					s = Subcriptores(id_c=r, id_u=u)
+					s.save()
 
-             return render(request, vervideo, {'q':q})
-        else :
-          return HttpResponseRedirect('/login')
-      else :
-        return render(request, vervideo, {'q':q})
+					return render(request, vervideo, {'q':q, 'form':form,'c':c})
+		else :
+			return HttpResponseRedirect('/login')
+	else :
+		return render(request, vervideo, {'q':q, 'form':form,'c':c})
 
 
 def viewchannel(request, channel):
